backend/requirements.txt
fastapi
uvicorn
openai
pandas
numpy
scipy
faker
openpyxl
python-dotenv
pydantic
python-multipart